somehow got the nasty profile fragment behave. usermanager now handles error roubustly and avoids unnecessary downloads

+++ app/src/main/java/com/pair/util/FileHelper.java
@@ -93,6 +93,9 @@ public class FileHelper {
 
     public static void save(File profilePicture, InputStream in) throws IOException {
         byte[] imageBytes = IOUtils.toByteArray(in);
+        if (profilePicture.exists()) {
+            return;
+        }
         FileUtils.writeByteArrayToFile(profilePicture, imageBytes);
     }
